+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 ## 2024-02-15
 ### Added
+  - added aliased action `:do_anything` (same as `:manage`), into Folio::Ability.
+  - added `user.currently_allowed_actions_with(subject, ability_class = nil)`
+
+## 2024-02-15
+### Added
 - added `Folio::TogglableFieldsComponent`
 
 ## 2024-02-08